Add files via upload
a denoising autoencoder is a modification of the original autoencoder in which instead of giving the original input we give a corrupted or noisy version of input to the encoder while decoder loss is calculated concerning original input only. This results in efficient learning of autoencoders and the risk of autoencoder becoming an identity function is significantly reduced. Using this an image processing model is developed and a dataset of CIFAR-10 is used for the training , testing and validation purpose.